Add device timeout monitoring
Part of the hardware interface requires that driver nodes report when there is
an issue by changing their state to OFF. Since this package is a wrapper around
the F1/10 vesc driver, it needs to listen for sensor updates. If the sensor
updates stop being published, that indicates something is wrong with the VESC,
and the driver status needs to be updated.

c1t_vesc_driver.h
* rename timeout variable to be more informative

c1t_vesc_driver.launch
* add a remap for the subscribing to the vesc node's sensor output topic
* add parameter file load for the timeout variable

c1t_vesc_driver.cpp
* add parameter initialization for sensor timeout threshold
* add subscriber initialization for vesc node sensor monitoring