Add Jetson TX2 Noetic Docker support for package
The package can now run inside a ROS Noetic Docker container on the Jetson TX2
platform. The Docker configuration is specific to the Jetson platform, but the
base image could be swapped out to run on a non-Jetson platform.

A note about the current configuration. The Jetson runs Ubuntu 18.04, which is
not supported for Noetic. Therefore, everything must be built from source.
The normal from-source process of using rosinstall_generator and rosdep will
not work here because the tools do not know how to deal with Noetic on 18.04.
Instead, all of the package's dependencies' repositories need to be
explicitly listed so that vcs can clone them. That is what the .repos files
are for.

Dockerfile
* add steps to build package for ROS Noetic from source

autoware-messages.repos
* add all repos necessary to build autoware_msgs package from source (assuming
  ros_base package sources are already installed)

carma-msgs.repos
* add all repos necessary to build cav_msgs package from source (assuming
  ros_base package sources are already installed)

carma-utils.repos
* add all repos necessary to build cav_driver_utils from source (assuming
  ros_base package sources are already installed)

docker/checkout.bash
* add script to download necessary repositories using vcs

docker/install.sh
* add script to build and install c1v_vesc_driver package from source

vesc.repos
* add all repos necessary to build vesc packages from source (assuming
  ros_base package sources are already installed)