Add carma user to dialout group in container
The VESC presents as a serial device, and non-root users can only access it if
they're either part of the dialout group or the udev rules have been configured
for anyone to read the device. To avoid relying on the latter scenario, the
carma user has been added to the dialout group.

Dockerfile
* add carma user to dialout group