3 - Change robot_active to not always be true
Setting the robot_active field to always true seemed to cause issues with the
guidance module transitioning from active to engaged. The robot_active field
has been set to equal robot_enabled. The VESC is always listening for incoming
commands, so it made sense to set it to active whenever it was also enabled.

c1t_vesc_driver.cpp
* update robot_active field to be equal to robot_enabled